References

Index

ChunkSplitters.chunks

ChunkSplitters.getchunk

ChunkSplitters

ChunkSplitters.chunks — Function

chunks(array::AbstractArray, nchunks::Int, type::Symbol=:batch)

This function returns an iterable object that will split the indices of array into to nchunks chunks. type can be :batch or :scatter. It can be used to directly iterate over the chunks of a collection in a multi-threaded manner.

Example

julia> using ChunkSplitters 

julia> x = rand(7);

julia> for i in chunks(x, 3, :batch)
           @show Threads.threadid(), collect(i)
       end
(Threads.threadid(), collect(i)) = (6, [1, 2, 3])
(Threads.threadid(), collect(i)) = (8, [4, 5])
(Threads.threadid(), collect(i)) = (7, [6, 7])

julia> for i in chunks(x, 3, :scatter)
           @show Threads.threadid(), collect(i)
       end
(Threads.threadid(), collect(i)) = (2, [1, 4, 7])
(Threads.threadid(), collect(i)) = (11, [2, 5])
(Threads.threadid(), collect(i)) = (3, [3, 6])

ChunkSplitters.getchunk — Function

getchunk(array::AbstractArray, ichunk::Int, nchunks::Int, type::Symbol=:batch)

Function that returns a range of indexes of array, given the number of chunks in which the array is to be split, nchunks, and the current chunk number ichunk.

If type == :batch, the ranges are consecutive. If type == :scatter, the range is scattered over the array.

Example

For example, if we have an array of 7 elements, and the work on the elements is divided into 3 chunks, we have (using the default type = :batch option):

julia> using ChunkSplitters

julia> x = rand(7);

julia> getchunk(x, 1, 3)
1:3

julia> getchunk(x, 2, 3)
4:5

julia> getchunk(x, 3, 3)
6:7

And using type = :scatter, we have:

julia> getchunk(x, 1, 3, :scatter)
1:3:7

julia> getchunk(x, 2, 3, :scatter)
2:3:5

julia> getchunk(x, 3, 3, :scatter)
3:3:6